dep-scanner/requirements.txt
gitpython>=3.1.0
pydantic>=2.0.0
pip-audit>=2.6.0
orchagent-sdk>=0.1.0
